post/requirements.txt
aioredis==2.0.1
anyio==3.7.0
async-timeout==4.0.2
fastapi==0.96.0
h11==0.14.0
h2==4.1.0
hiredis==2.2.3
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
priority==2.0.0
psycopg2-binary==2.9.6
pydantic==1.10.8
python-dotenv==1.0.0
redis==4.6.0
sniffio==1.3.0
SQLAlchemy==2.0.17
sqlalchemy-cockroachdb==2.0.1
SQLAlchemy-serializer==1.4.1
starlette==0.27.0
toml==0.10.2
typing_extensions==4.6.3
wsproto==1.2.0
